Delete obsolete files
