--- bookstore/src/components/Categories.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Categories = () => (
-  <button type="button">Check status</button>
-);
-
-export default Categories;
