--- bookstore/src/components/Form.js
@@ -1,14 +1,0 @@
-import React from 'react';
-
-const Form = () => (
-  <div>
-    <h2>ADD NEW BOOK</h2>
-    <form action="post">
-      <input type="text" placeholder="Book title" />
-      <input type="text" placeholder="Author" />
-      <button type="button">ADD BOOK</button>
-    </form>
-  </div>
-);
-
-export default Form;
