--- bookstore/src/index.css
@@ -1,3 +1,0 @@
-.user-icons {
-  font-size: 50px;
-}
