+++ src/App.js
@@ -1,0 +1,19 @@
+import './index.css';
+import { Routes, Route } from 'react-router-dom';
+import Navbar from './components/Navbar';
+import Books from './components/Books';
+import Categories from './components/Categories';
+
+function App() {
+  return (
+    <div className="App">
+      <Navbar />
+      <Routes>
+        <Route path="/" element={<Books />} />
+        <Route path="/Categories" element={<Categories />} />
+      </Routes>
+    </div>
+  );
+}
+
+export default App;
