+++ src/components/Book.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+const Book = ({ title, author }) => (
+  <div>
+    <h1>{title}</h1>
+    <p>{author}</p>
+    <button type="button">Remove</button>
+  </div>
+);
+
+Book.propTypes = {
+  title: PropTypes.string.isRequired,
+  author: PropTypes.string.isRequired,
+};
+export default Book;
