+++ src/index.css
@@ -1,0 +1,3 @@
+.user-icons {
+  font-size: 50px;
+}
